refactor(components/molecule/radioButtonGroup): replace cloneElement for injector
