+++ components/molecule/radioButtonGroup/demo/articles/ArticleMoleculeRadioButtonField.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 import {Article, H2, Paragraph} from '@s-ui/documentation-library'
 import MoleculeRadioButtonField from '@s-ui/react-molecule-radio-button-field'
 
-import MoleculeRadioButtonGroup from '../src/index.js'
-import CustomLabel from './components/CustomLabel.js'
+import MoleculeRadioButtonGroup from '../../src/index.js'
+import CustomLabel from '../components/CustomLabel.js'
 
 const ArticleMoleculeRadioButtonField = ({className}) => {
   return (
