+++ components/molecule/radioButtonGroup/package.json
@@ -9,7 +9,8 @@
     "build:styles": "cpx './src/**/*.scss' ./lib"
   },
   "dependencies": {
-    "@s-ui/component-dependencies": "1"
+    "@s-ui/component-dependencies": "1",
+    "@s-ui/react-primitive-injector": "1"
   },
   "keywords": [],
   "author": "",
